requirements.txt
altair==4.2.2
contractions==0.1.73
demoji==1.1.0
matplotlib==3.5.1
nltk==3.8.1
pandas==1.5.2
pyspellchecker==0.7.1
scikit_learn==1.2.2
streamlit==1.20.0
wordcloud==1.8.2.2
